accounts/urls.py
from django.urls import path
from .views import *
from rest_framework_simplejwt.views import TokenRefreshView

app_name = 'accounts'

urlpatterns = [
    path("signup/", SignupView.as_view(), name='sign-up'),
    path("login/", LoginView.as_view(), name="login"),
    path("logout/", LogoutView.as_view(), name="logout"),
]